Renamed IViewModFilesViewModel to IModFilesViewModel
